requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
together
langchain
chromadb
sentence-transformers


